0.3 release
A pretty substantial re-write. Sample the Vcc value to try and improve the A/D accuracy. Add a lot more comments to explain better what's going on. This version has actually made boards, so the code can be considered ready for prime time. It's still the hardware on my breadboard, so I'm still using the AD595 instead of the 8495. That should not wind up making a lot of difference except for the TEMP_SCALING_FACTOR value.